NexusEnrichmentAgent/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "NexusEnrichmentAgent"
version = "1.0.0"
description = "Multi-agent enrichment system for compliance control mapping"
requires-python = ">=3.9"
dependencies = [
    "strands-agents",
    "boto3>=1.26.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black",
    "isort>=5.0",
    "mypy>=1.0",
    "flake8>=6.0",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["nexus_enrichment_agent"]

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
show_error_codes = true
pretty = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["test"]
asyncio_mode = "auto"
addopts = "-v --cov=nexus_enrichment_agent --cov-report=term-missing"
